requirements.txt
pandas==1.5.0
requests==2.28.1
apache-airflow-providers-mysql